% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CESKalman_Bootstrap.R
\name{CESKalman_Bootstrap}
\alias{CESKalman_Bootstrap}
\title{Function for bootstrapping confidence intervals of the elasticity and adjustment parameter}
\usage{
CESKalman_Bootstrap(Estimation,grid.param_init=c(-9,-1,3),ndraw=1000,print_results=TRUE,
                           cVal_Auto=0.1,cVal_NIS=0.1)
}
\arguments{
\item{Estimation}{An object of class CESKalman}

\item{grid.param_init}{A vector of initial Parameter values of the observation variance to loop over (see details)}

\item{ndraw}{The number of draws in the bootstrapping procedure, 1000 is default.}

\item{cVal_NIS}{Critical value for the NIS test}

\item{print_results}{Should results be printed while bootstrapping?}

\item{cVal_auto}{Critical value for the Breusch-Godfrey test for autocorrelation}
}
\value{
Returns a matrix of dimension ndrawX3. First column is the draws of sigma, second is the draws of alpha and third is the likelihood value.
}
\description{
The function uses a recursive-design bootstrapping procedure to produce confidence intervals of
the elasticity of substitution and adjustment parameter. The CESKalman_Estimation function is applied and
estimated for every new draw.
}
\details{
This function takes an object of class CESKalman as input and calls the function CESKalman_Estimation for every new draw. The parameter estimates
from CESKalman are used as initial values of sigma and alpha. Also, the number of lags and lambda are determined
by the CESKalman function.

grid.param_init can be a vector of any length. Values to loop over for the variance of error term in observation equation, \eqn{\epsilon_t}, in the numerical optimization. When lambda is FALSE, differen values
for the signal-to-noise ratio in the range from 10-1000 found to be optimal in Kronborg et al (2019) are tried. The one that maximizes the likelihood is chosen.

Only draws with no autocorrelation in the estimated model and the NIS test within the limits specified in cVal_NIS is accepted.
}
\examples{

## First, data is loaded with the Load_Data function (or any other data set)
data = Load_Data(Country="USA",tstart=1970,tend=2017)

data = cbind(data[,"q"],data[,"w"],data[,"K"],data[,"L"])

## Next, the CESKalman function is called
Kalman = CESKalman(data=data,grid.lambda=c(10,500,20), lambda_est_freely=TRUE)

## Bootstrapping confidence intervals
Bootstrap = CESKalman_Bootstrap(Estimation=Kalman)

}
\references{
Kronborg et al (2019) and Petris et al (2010) and Kastrup et al (2021)
}
\author{
Christian Sandholm Kastrup <CST@dreamgruppen.dk>, Anders Farver Kronborg <ANK@dreamgruppen.dk> and Peter Philip Stephensen <PSP@dreamgruppen.dk>
}
