% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CESKalman_Estimation.R
\name{CESKalman_Estimation}
\alias{CESKalman_Estimation}
\title{Function for estimating a CES production function with the Kalman filter.}
\usage{
CESKalman_Estimation(Y,X,grid.param_init=c(-9,-1,5),nlags,lambda,

Leontief=FALSE,alpha_init,sigma_init,cVal_NIS=0.10)
}
\arguments{
\item{Y}{First-difference of log relative budget shares}

\item{X}{A matrix of all explanatory variables to be used (See details)}

\item{grid.param_init}{A string of length 3. First and second argument are the, respectively, lower and upper bound of the grid of the initial variance of the error term used in the dlmMLE function. Third is the step size}

\item{nlags}{Number of lags used of relative prices and relative budget shares. One of 0,1,2}

\item{lambda}{The inverse of the signal-to-noise ratio. Can be any positive value. If set to FALSE (the default) it is estimated}

\item{Leontief}{TRUE or FALSE (default). Should the elasticity of substitution be fixed at zero? If so, set =TRUE}

\item{alpha_init}{Initial value for the adjustment parameter, \eqn{\alpha}}

\item{sigma_init}{Initial value for the long run elasticity of substitution, \eqn{\sigma}}

\item{cVal_NIS}{Critical value used to construct the confidence bands of the NIS test. Default is 0.10}
}
\value{
Returns a list with the following objects:

\strong{Smooth:} The list returned from the dlmSmooth function.
\strong{Gamma:} The relative augmenting technologies in logarithms (input 1 relative to input 2).
\strong{residuals:} The residuals from the observation equation.
\strong{sigma:} The long run elasticity of substitution, \eqn{\sigma}.
\strong{alpha:} The adjustment parameter, \eqn{\alpha}.
\strong{LV:} The likelihood value.
\strong{est.lambda:} The estimated value of lambda. (only estimated if lambda was specified as FALSE, else just the chosen value).
\strong{BG_test:} p-value from the Breusch Godfrey test for autocorrelation from function bgtest.
\strong{BP_test:} p-value from the Breusch Pagan test for heteroscedasticity from function bptest.
\strong{JB_test:} p-value from the Jarque Bera test for normality from function jbtest.
\strong{NIS_test:} Value, lower and upper confidence bands of the Normalized Innovations Squared test.
}
\description{
This function first estimates the variances with the dlmMLE function from dlm package. Second, the
dlmSmooth function is applied to estimate the state parameters. Third, several statistical tests
are performed using the bgtest and bptest from lmtest package and the jb.norm.test from package normtest. Also, it calculates the normalized innovations squared.
}
\details{
This function first estimates the variances with the dlmMLE function from dlm package. Second, the
dlmSmooth function is applied to estimate the state parameters. Third, several statistical tests
are performed using the bgtest and bptest from lmtest package and the jaruqe.bera.test from package normtest. Also, it calculates the normalized innovations squared.

\strong{NOTE:} This function can be used on its own, but we encourage to use the CESKalman instead as it is more robust.

The estimated function is the Error-Correction model from Kronborg et al (2019):
\eqn{\Delta s_{t}=\alpha(s_{t-1}-(1-\sigma)p_{t-1}-\mu_{t-1})+\sum_{i=0}^{nlags}\kappa_{i}\Delta p_{t-i}+\sum_{i=1}^{nlags}\gamma_{i}\Delta s_{t-i}+\epsilon_t}.
\eqn{s_t} is the relative budget shares in logs (expenditure on factor 1 relative to factor 2), \eqn{p_t} is the relative prices in logs and \eqn{\mu_t=(\sigma-1)log(\Gamma_t)} where \eqn{\Gamma_t} is the relative
augmenting technical change. The process of \eqn{\mu_t}, the state variable, is an I(2) process: \eqn{\Delta \mu_t=\Delta \mu_{t-1}+\eta_t}.

X should be a matrix with all the explanatory variables to be used (all in logs). The dimension is TxK, where T
is time periods and K are the explanatory variables (not counting technical change as one). K=3 when nlags=0, =5 when nlags=1 and =7 when nlags=2. The column ordering should be:
1: lags of relative budget shares in log-levels. 2: lags of the relative prices in log-levels. 3: First-difference of log relative prices.
4: First-differences of log relative prices lagged one period. 5: First-differences of log relative budget shares lagged one period.
6: First-differences of log relative prices lagged two periods. 7: First-differences of log relative budget shares lagged two periods.
When using the CESKalman, CESKalman_Estimation or CESKalman_Bootstrap functions, data will be ordered in this way automatically.

grid.param_init can be a vector of any length. It is the values to loop over for the variance of the error term in observation equation, \eqn{\epsilon_t}, in the numerical optimization. When lambda is set to NA, different values
for the signal-to-noise ratio in the range from 10-1000 found to be optimal in Kronborg et al (2019) are tried. The one that maximizes the likelihood is chosen.
}
\seealso{
See dlm package for a decription of dlm objects, in particular the dlmSmooth function. See Kronborg et al (2019) for a description of the methodology.
}
\author{
Christian Sandholm Kastrup <CST@dreamgruppen.dk> , Anders Farver Kronborg <ANK@dreamgruppen.dk> and Peter Philip Stephensen <psp@dreamgruppen.dk>
}
