% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.CESKalman.R
\name{plot.CESKalman}
\alias{plot.CESKalman}
\title{Function for plotting the model fit and trend/price decomposition of a CESKalman object}
\usage{
plot.CESKalman(Kalman,t0=1,tEnd=nrow(Kalman$data))
}
\arguments{
\item{Kalman}{an object of class CESKalman returned from the CESKalman function}

\item{t0}{Start date. Only available for yearly data, else it is an index}

\item{tEnd}{End date. Only available for yearly data, else it is an index}
}
\description{
The first plot contains the model fit of the relative quantities. The second is the process of relative augmenting technical change.
The third plot is the demeaned-data series applied in estimation.
}
\details{
\strong{NOTE:}
As we are primarily interested to explain the model fit of the relative quantities (e.g. capital relative to labor), the estimated equation is first translated into quantities, denoted \eqn{x_t}:
\eqn{\Delta x_{t}=\alpha(x_{t-1}+\sigma p_{t-1}-\mu_{t-1})+(\kappa_{i}-1)\Delta p_{t}+\epsilon_t}. \eqn{x_t} is the relative quantities, 
\eqn{p_t} the relative prices (e.g. the user cost relative to the wage) and \eqn{\mu_{t} \equiv (\sigma-1)log(\Gamma_t)} with \eqn{\Gamma_t} 
being the process of relative augmenting technical change (e.g. capital augmenting techncial change relative to labor augmenting technical change).

The first graph shows the values of \eqn{x_{t}}, the fitted values,\eqn{\hat{x_{t}}=x_{t}-\epsilon_t} and the residuals \eqn{\epsilon_t}. It is used to
evaluate the fit of the model. 

The second graph shows the process of relative augmenting technical change, i.e. \eqn{log(\Gamma_t)} and shows the direction of technical change. Of particular interest
is if the direction has changed throught the sample and features so-called "medium run" fluctuations.   

The last graph shows the data series of \eqn{x_t-\bar{x}} and \eqn{-(p_t-\bar{p})} with \eqn{\bar{x}} and \eqn{\bar{p}} denoting the mean over time. 
Note that the relative price is plotted on the right axis and with a negative sign in front. If the two resulting series are possitively correlated, it implies
that the long run elasticity is expected to be positive.
}
\examples{
## First, data is loaded with the Load_Data function (or any other data set)
data = Load_Data(Country="USA",tstart=1970,tend=2017)

data = cbind(data[,"q"],data[,"w"],data[,"K"],data[,"L"])

## The second step is a call to the CESKalman function
Kalman = CESKalman(data,grid.lambda=c(10,500,20),lambda_est_freely = T)

## Lastly we can plot the fit of the model:
plot(Kalman)

}
\references{
Kronborg et al (2019) and Kastrup et al (2021)
}
\author{
Christian Sandholm Kastrup <CST@dreamgruppen.dk>, Anders Farver Kronborg <ANK@dreamgruppen.dk> and Peter Philip Stephensen <PSP@dreamgruppen.dk>
}
